wlsunset

+++ packages/sway.sh
@@ -6,7 +6,7 @@ install() {
     install_apt swaylock
     install_apt swayidle
 
-    # When updating to Ubuntu 23+ this can be replaced by install_apt sway-notification-center
+    # When updating to Ubuntu 23+ this (deps below, +wayland-protocols & mako) can be replaced by install_apt sway-notification-center
     install_apt meson
     install_apt libwayland-dev
     install_apt libpango1.0-dev
@@ -25,13 +25,26 @@ install() {
     execute cd ..
     execute rm -rf wayland-protocols-1.32.tar.xz wayland-protocols-1.32
 
-    execute rm -rf mako
-    execute git clone https://github.com/emersion/mako.git
-    execute cd mako
-    execute meson build
-    execute ninja -C build
-    execute sudo ninja -C build install
-    execute cd ..
-    execute rm -rf mako
+    if ! command -v mako &> /dev/null; then
+      execute rm -rf mako
+      execute git clone https://github.com/emersion/mako.git
+      execute cd mako
+      execute meson build
+      execute ninja -C build
+      execute sudo ninja -C build install
+      execute cd ..
+      execute rm -rf mako
+    fi
+
+    if ! command -v wlsunset &> /dev/null; then
+      execute rm -rf wlsunset
+      execute git clone https://github.com/kennylevinsen/wlsunset.git
+      execute cd wlsunset
+      execute meson build
+      execute ninja -C build
+      execute sudo ninja -C build install
+      execute cd ..
+      execute rm -rf wlsunset
+    fi
   fi
 }
